added advanced cluster file
